+++ src/lib.rs
@@ -15,10 +15,51 @@ macro_rules! impl_extra {
 
         impl_extra!(@all, $owned, $ref);
     };
+    (no_arb, $owned:path, $ref:path) => {
+        impl<'a> From<&'a String> for &'a $ref {
+            fn from(string: &'a String) -> Self {
+                <$ref>::from_str(string.as_str())
+            }
+        }
+
+        impl_extra!(@all, $owned, $ref);
+    };
     ($owned:path, $ref:path) => {
         impl<'a> From<&'a String> for &'a $ref {
             fn from(string: &'a String) -> Self {
                 <$ref>::from_str(string.as_str())
+            }
+        }
+
+        #[cfg(feature = "arbitrary")]
+        impl<'a> arbitrary::Arbitrary<'a> for &'a $ref {
+            fn arbitrary(u: &mut arbitrary::Unstructured<'a>) -> Result<Self, arbitrary::Error> {
+                <&str as arbitrary::Arbitrary>::arbitrary(u).map(Into::into)
+            }
+
+            fn arbitrary_take_rest(u: arbitrary::Unstructured<'a>) -> Result<Self, arbitrary::Error> {
+                <&str as arbitrary::Arbitrary>::arbitrary_take_rest(u).map(Into::into)
+            }
+
+            #[inline]
+            fn size_hint(depth: usize) -> (usize, Option<usize>) {
+                <&str as arbitrary::Arbitrary>::size_hint(depth)
+            }
+        }
+
+        #[cfg(feature = "arbitrary")]
+        impl<'a> arbitrary::Arbitrary<'a> for $owned {
+            fn arbitrary(u: &mut arbitrary::Unstructured<'a>) -> Result<Self, arbitrary::Error> {
+                <&$ref as arbitrary::Arbitrary>::arbitrary(u).map(Into::into)
+            }
+
+            fn arbitrary_take_rest(u: arbitrary::Unstructured<'a>) -> Result<Self, arbitrary::Error> {
+                <&$ref as arbitrary::Arbitrary>::arbitrary_take_rest(u).map(Into::into)
+            }
+
+            #[inline]
+            fn size_hint(depth: usize) -> (usize, Option<usize>) {
+                <&$ref as arbitrary::Arbitrary>::size_hint(depth)
             }
         }
 
@@ -30,6 +71,22 @@ macro_rules! impl_extra {
             #[doc = concat!("[`Cow<'_, ", stringify!($ref), ">`](std::borrow::Cow::Borrowed)")]
             pub fn as_cow<'a>(&'a self) -> ::std::borrow::Cow<'a, $ref> {
                 self.into()
+            }
+        }
+
+        #[cfg(feature = "zerofrom")]
+        impl<'zf> zerofrom::ZeroFrom<'zf, $ref> for &'zf $ref {
+            #[inline]
+            fn zero_from(other: &'zf $ref) -> Self {
+                other
+            }
+        }
+
+        #[cfg(feature = "zerofrom")]
+        impl<'zf> zerofrom::ZeroFrom<'zf, $owned> for &'zf $ref {
+            #[inline]
+            fn zero_from(other: &'zf $owned) -> Self {
+                other
             }
         }
 
@@ -62,6 +119,24 @@ macro_rules! impl_extra {
                 ::std::borrow::Cow::Borrowed(self.as_ref())
             }
         }
+
+        const _: () = {
+            #[cfg(feature = "arbitrary")]
+            fn assert_arbitrary() {
+                fn assert<'a, T: arbitrary::Arbitrary<'a>>() {}
+                // XXX: Not asserting arbitrary is implemented for borrowed, this is because a validated type might need owned data.
+                // assert::<&$ref>();
+                assert::<$owned>();
+            }
+
+            #[cfg(feature = "zerofrom")]
+            fn assert_zerofrom() {
+                fn assert_borrowed<'zf, T: zerofrom::ZeroFrom<'zf, $ref>>() {}
+                fn assert_owned<'zf, T: zerofrom::ZeroFrom<'zf, $owned>>() {}
+                assert_borrowed::<&$ref>();
+                assert_owned::<&$ref>();
+            }
+        };
     };
 }